+++ CHANGELOG.md
@@ -6,15 +6,19 @@
 - Filtering issues table
 - Logging and a log viewer in settings
 
-### Completed
+### Added
 
 - Display version in settings, check git tags for updates
 
 # 1.1.0
 
+### Added
+
 - Dashboard widget displaying time tracking from the previous working day
 - Kanban-style to do list with ability to link jira issues
 
 # 1.0.0
+
+### Added
 
 - Initial version ported from WPF app